+++ lang/total.php
@@ -3656,16 +3656,16 @@ Ein anderer Begriff für Niveaustufen ist Taxonomien - z.B. kann die Bloomsche T
         'Preselected for deletion because this subject is disabled in Komet',
     ],
     'preselect_delete_subject_because_it_was_not_imported_in_last_import' => [
-        'Vorausgewählt für Löschung, da dieses Fach auf Komet nicht mehr vorhanden ist, aber in dieser Moodle-Installation noch in Kursen verwendet wird, oder Bewertungen hat, oder Inhalte anderer Quellen beinhaltet.',
+        'Vorausgewählt für Löschung, da dieses Fach im letzten Import nicht mehr vorhanden war, aber in dieser Moodle-Installation noch in Kursen verwendet wird, oder Bewertungen hat, oder Inhalte anderer Quellen beinhaltet.',
         'Preselected for deletion because this subject did not exist in the previous import task BUT is being used in courses or has gradings or has has content from another source.',
     ],
-    'sync_all_grids_with_komet' => [
-        'Synchronisiere alle Raster mit Komet',
-        'Sync all grids with Komet',
+    'delete_grids_missing_from_xmlserverurl' => [
+        'Daten bei Synchronisation mit xmlserverurl löschen',
+        'Delete data during synchronization with xmlserverurl',
     ],
-    'sync_all_grids_with_komet_description' => [
-        'Synchronisiere alle Raster mit Komet. Alle Raster, die in Komet nicht mehr vorhanden sind, werden gelöscht, es sei denn, sie sind bereits in Verwendung.',
-        'Keep only the Komet grids and competencies. Delete everything else unless a grid or competency is already in use',
+    'delete_grids_missing_from_xmlserverurl_description' => [
+        'Alle Raster, die in Komet (xmlserverurl) nicht vorhanden sind, werden gelöscht, es sei denn, sie sind bereits in Verwendung.',
+        'All grids that no longer exist in Komet (xmlserverurl) will be deleted unless they are already in use.',
     ],
     'no_examples_in_this_grid' => [
         'In diesem Raster gibt es keine Lernmaterialien',